Minor optimization and refactoring for comps, alchemy, and player. Ready to merge.
