+++ client/src/components/CommonPlayerCreaturePanel.jsx
