+++ client/src/pages/Lobby.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import { useState, useEffect } from "react";
 import "../App.scss";
 import "./Lobby.css";
 import Userfront from "@userfront/core";
@@ -6,6 +6,7 @@ import { useNavigate } from "react-router-dom";
 import { getUser } from "../services/userServices";
 import GameNav from "../layouts/GameNav";
 import OptionsMenu from "../layouts/OptionsMenu";
+import PlayerPanel from "../components/PlayerPanel";
 import MultiPlayerGameMenu from "../layouts/MultiPlayerGameMenu";
 import MultiPlayerCreature from "../components/MultiPlayerCreature";
 import MultiPlayerEnemyCreature from "../components/MultiPlayerEnemyCreature";
@@ -111,7 +112,7 @@ function Lobby() {
       }
     };
     // retrieves user data and updates player state
-    const loadAsyncDataPlayer = async () => {
+    const initAsyncDataPlayer = async () => {
       try {
         const { data } = await getUser();
         setPlayer(data);
@@ -120,7 +121,7 @@ function Lobby() {
       }
     };
     initAsyncDataLobby();
-    loadAsyncDataPlayer();
+    initAsyncDataPlayer();
   }, []);
 
   useEffect(() => {
@@ -206,48 +207,6 @@ function Lobby() {
     dispatch,
   ]);
 
-  // renders player details panel
-  const getPlayer = () => {
-    return (
-      <>
-        <div className="color_white">
-          <img
-            src={player.avatarPath}
-            alt={player.name}
-            className="player_avatar"
-            width="96"
-            height="96"
-          />
-          <h4>{player.name}</h4>
-          <h5>
-            Lvl. {Math.floor(Math.sqrt(player.experience) * 0.25)} |{" "}
-            {player.experience.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",")}{" "}
-            XP
-            <div className="progress_bar_container">
-              <div
-                className="progress_bar"
-                style={{
-                  width:
-                    (
-                      Math.sqrt(player.experience) * 0.25 -
-                      Math.floor(Math.sqrt(player.experience) * 0.25)
-                    )
-                      .toFixed(2)
-                      .replace("0.", "") + "%",
-                }}
-              />
-            </div>
-          </h5>
-          <h5>
-            Drachmas:{" "}
-            {player.drachmas.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",")}{" "}
-            {"\u25C9"}
-          </h5>
-        </div>
-      </>
-    );
-  };
-
   // retrieves user data and updates player state
   const loadAsyncDataPlayer = async () => {
     try {
@@ -321,7 +280,7 @@ function Lobby() {
         </header>
 
         <main className="lobby1_game_section">
-          {getPlayer()}
+          <PlayerPanel player={player} />
 
           {/* displays other menus and creatures if options menu isnt being used */}
           {Object.values(optionsMenuStatus).every(
